+++ sig/cpp_source.rbs
@@ -20,7 +20,7 @@ module Wrapture
     def self.format_class_declaration: (Wrapture::CppSource::CppDeclaration) -> Enumerable[String]
     def self.format_class_definition: (Wrapture::CppSource::CppClass) -> Enumerable[String]
     def self.format_declaration: (Wrapture::CppSource::CppDeclaration) -> Enumerable[String]
-    def self.format_method_declaration: (Wrapture::CppSource::CppMethod) -> Enumerable[String]
+    def self.format_member_function_declaration: (Wrapture::CppSource::CppFunction) -> Enumerable[String]
     def self.format_tree: (Enumerable[untyped]) -> Enumerable[String]
   end
 end
