--- sig/cpp_source/cpp_method.rbs
@@ -1,35 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-# Copyright 2025 Joel E. Anderson
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-module Wrapture
-  module CppSource
-    class CppMethod
-      @accessibility: Symbol
-      @name: String
-      @return_type: CppType
-      @tree: Array[untyped]
-
-      attr_accessor accessibility: Symbol
-      attr_accessor attributes: Array[String]
-      attr_reader name: String
-      attr_reader params: Array[untyped]
-      attr_accessor return_type: CppType
-      attr_reader tree: Array[untyped]
-
-      def initialize: (String) -> void
-    end
-  end
-end
